update examples and references
