+++ examples/list-incomplete-uploads.mjs
@@ -1,5 +1,5 @@
 /*
- * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2021 MinIO, Inc.
+ * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2015 MinIO, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-// Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname
-// and my-objectname are dummy values, please replace them with original values.
+// Note: YOUR-ACCESSKEYID, YOUR-SECRETACCESSKEY, my-bucketname and my-prefixname
+// are dummy values, please replace them with original values.
 
 import * as Minio from 'minio'
 
@@ -25,15 +25,11 @@ const s3Client = new Minio.Client({
   secretKey: 'YOUR-SECRETACCESSKEY',
 })
 
-// Enable S3 transfer accelerate endpoint.
-s3Client.SetS3TransferAccelerate('s3-accelerate.amazonaws.com')
-
-// Upload a buffer
-const buf = new Buffer(10)
-buf.fill('a')
-s3Client.putObject('my-bucketname', 'my-objectname2', buf, 'application/octet-stream', function (e) {
-  if (e) {
-    return console.log(e.message)
-  }
-  console.log('Success')
+// List all incomplete uploads in bucket my-bucketname that begin with my-prefixname.
+const incompleteObjectsStream = s3Client.listIncompleteUploads('my-bucketname', 'my-prefixname', true)
+incompleteObjectsStream.on('data', function (obj) {
+  console.log(obj)
+})
+incompleteObjectsStream.on('error', function (e) {
+  console.log(e)
 })
